bug fix with giza_rgb_from_table_float
git-svn-id: https://svn.code.sf.net/p/giza/code@290 952f94de-64c9-43ce-8081-c12abd9885bf

+++ include/giza.h
@@ -108,7 +108,7 @@ void giza_save_colour_table (void);
 void giza_restore_colour_table (void);
 void giza_rgb_from_table (double pos, double *red, double *green,
 			  double *blue);
-void giza_rgb_from_table_float (double pos, float *red, float *green,
+void giza_rgb_from_table_float (float pos, float *red, float *green,
 				float *blue);
 
 void giza_contour (int sizex, int sizey, const double* data,
